+++ Fabric/src/main/java/fuzs/puzzleslib/impl/event/FabricEventInvokerRegistryImpl.java
@@ -1,20 +1,33 @@
 package fuzs.puzzleslib.impl.event;
 
 import com.google.common.collect.MapMaker;
+import fuzs.puzzleslib.api.core.v1.ModLoaderEnvironment;
 import fuzs.puzzleslib.api.event.v1.*;
 import fuzs.puzzleslib.api.event.v1.core.EventInvoker;
 import fuzs.puzzleslib.api.event.v1.core.EventPhase;
 import fuzs.puzzleslib.api.event.v1.core.FabricEventInvokerRegistry;
+import fuzs.puzzleslib.api.event.v1.data.DefaultedValue;
 import fuzs.puzzleslib.api.event.v1.entity.living.LivingExperienceDropCallback;
 import fuzs.puzzleslib.api.event.v1.entity.living.LivingFallCallback;
 import fuzs.puzzleslib.api.event.v1.entity.player.BonemealCallback;
 import fuzs.puzzleslib.api.event.v1.entity.player.PlayerInteractEvents;
 import fuzs.puzzleslib.api.event.v1.entity.player.PlayerXpEvents;
 import fuzs.puzzleslib.api.event.v1.world.BlockEvents;
+import fuzs.puzzleslib.impl.client.event.FabricClientEventInvokers;
+import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
 import net.fabricmc.fabric.api.event.Event;
 import net.fabricmc.fabric.api.event.player.UseItemCallback;
+import net.fabricmc.fabric.api.loot.v2.LootTableEvents;
+import net.fabricmc.fabric.api.loot.v2.LootTableSource;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.server.packs.resources.ResourceManager;
 import net.minecraft.world.InteractionResultHolder;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.storage.loot.LootTable;
+import net.minecraft.world.level.storage.loot.LootTables;
+import net.minecraft.world.level.storage.loot.providers.number.ConstantValue;
+import org.apache.commons.lang3.mutable.MutableBoolean;
+import org.apache.commons.lang3.mutable.MutableInt;
 
 import java.util.*;
 import java.util.function.Function;
@@ -34,6 +47,41 @@ public class FabricEventInvokerRegistryImpl implements FabricEventInvokerRegistr
         INSTANCE.register(PlayerTickEvents.Start.class, FabricEvents.PLAYER_TICK_START);
         INSTANCE.register(PlayerTickEvents.End.class, FabricEvents.PLAYER_TICK_END);
         INSTANCE.register(LivingFallCallback.class, FabricLivingEvents.LIVING_FALL);
+        INSTANCE.register(RegisterCommandsCallback.class, CommandRegistrationCallback.EVENT, callback -> {
+            return callback::onRegisterCommands;
+        });
+        INSTANCE.register(LootTableLoadEvents.Replace.class, LootTableEvents.REPLACE, callback -> {
+            return (ResourceManager resourceManager, LootTables lootManager, ResourceLocation id, LootTable original, LootTableSource source) -> {
+                // keep this the same as Forge where editing data pack specified loot tables is not supported
+                if (source == LootTableSource.DATA_PACK) return null;
+                DefaultedValue<LootTable> lootTable = DefaultedValue.fromValue(original);
+                callback.onReplaceLootTable(lootManager, id, lootTable);
+                // returning null will prompt no change
+                return lootTable.getAsOptional().orElse(null);
+            };
+        });
+        INSTANCE.register(LootTableLoadEvents.Modify.class, LootTableEvents.MODIFY, callback -> {
+            return (ResourceManager resourceManager, LootTables lootManager, ResourceLocation id, LootTable.Builder tableBuilder, LootTableSource source) -> {
+                // keep this the same as Forge where editing data pack specified loot tables is not supported
+                if (source == LootTableSource.DATA_PACK) return;
+                callback.onModifyLootTable(lootManager, id, tableBuilder::pool, index -> {
+                    MutableInt currentIndex = new MutableInt();
+                    MutableBoolean result = new MutableBoolean();
+                    tableBuilder.modifyPools(builder -> {
+                        if (index == currentIndex.getAndIncrement()) {
+                            // there is no way in Fabric Api to remove loot pools, but this seems to work for disabling at least
+                            builder.setRolls(ConstantValue.exactly(0.0F));
+                            builder.setBonusRolls(ConstantValue.exactly(0.0F));
+                            result.setTrue();
+                        }
+                    });
+                    return result.booleanValue();
+                });
+            };
+        });
+        if (ModLoaderEnvironment.INSTANCE.isClient()) {
+            FabricClientEventInvokers.register();
+        }
     }
 
     @SuppressWarnings("unchecked")